+++ src/main/java/com/x1/groo/diary/repository/DiaryEmotionRepository.java
@@ -3,4 +3,4 @@ package com.x1.groo.diary.repository;
 import com.x1.groo.diary.entity.DiaryEmotion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiaryEmotionRepository extends JpaRepository<DiaryEmotion, Long> { }
+public interface DiaryEmotionRepository extends JpaRepository<DiaryEmotion, Integer> { }
