+++ src/main/java/com/x1/groo/diary/repository/DiaryRepository.java
@@ -3,4 +3,4 @@ package com.x1.groo.diary.repository;
 import com.x1.groo.diary.entity.Diary;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiaryRepository extends JpaRepository<Diary, Long> { }
+public interface DiaryRepository extends JpaRepository<Diary, Integer> { }
