+++ src/main/java/com/x1/groo/diary/service/DiaryService.java
@@ -6,5 +6,5 @@ import com.x1.groo.diary.dto.DiaryRequestDTO;
  * 사용자 ID를 파라미터로 받아 일기를 생성
  */
 public interface DiaryService {
-    void createDiary(DiaryRequestDTO request, Long userId);
+    void createDiary(DiaryRequestDTO request, int userId);
 }
